Migrate the Language Restriction feature to SiteLanguage
TYPO3 introduced the SiteConfiguration together with the SiteLanguage feature and moves away from using `sys_language` table.

The impact: When creating a new language this won't be stored in `sys_language` anymore but in the SiteConfiguration in a file on the drive. As the extension uses the table to populate its language selectors, this will be empty or incomplete.

This PR will change the TCA in a way to get the languages from the SIteConfiguration and using the languageID to generate the RestrictionCollection.