[PERFORMANCE] Improving foreach loop
Remove the `empty()` check before the `foreach` as this is not necessary; an foreach won't loop through an empty array anyway.

Second, filter the TCA tables to only those which should be exported outside of the loop. Therefore we can remove the check inside the loop if the current active table in the loop should be exported. That improves performance and avoid getting slower in case the user adds more tables to TCA.