[BUGFIX] Fix the checkboxes at export views
In the export view (Excel/XML) there are some checkboxes which controls the export behaviour. None of them did work so far. Some are not reloading the page when clicked, do not remember its state and after all are wrongly evaluated in the backend. As a result they have no function at all.

This PR fixes:
- the checkboxes and language dropdown to reload the page and get evaluated in the backend
- checking for UTF8
- skipping the XML-Check

What not works yet:
- Forcing a different target language
- Checking for existing translations