[BUGFIX] Fix the checkboxes at import view
In the XML import view there are some checkboxes which controls the import behaviour. None of them did work so far as not reloading the page when clicked and do not remember its state and after all are wrongly evaluated in the backend. As a result they have no function at all.

This PR fixes:
- the checkboxes reload the page and get evaluated in the backend
- delete localizations before import

What not works yet:
- Import as default language