[BUGFIX] Adjust registration of columns to avoid cache issues
This commit adjust the issue of missing columns threw database updates performed by the TYPO3 console.

Due to the caching mechanism of TYPO3 12, the previous implementation only worked in case the administrator perform a database compare within the backend.

Adding a language restriction to a table is now done in two steps:

1. register the field for a table
2. configure the field (TCA)

The registration of a language restriction is done within the file 'ext_localconf.php'.

This is done by calling 'LanguageRestrictionRegistry::getInstance()->registerField()' within the extension, table and field name.

Configuring of a field happens within the TCA override.

Refs: #91