[BUGFIX] TCEMAIN: Missing or empty pid leads to an exception
The pid of some tables or data is empty or 0 by default.
The table sys_file_metadata currently does not contain a pid.

The empty pid leads to an exception which abort the whole process and will be visible within the backend.

In these special cases TYPO3 offers a method to retrieve the system languages (by using all site configurations).

Refs: #79